Add implementation of draft reply and react specs
For reply, we validate that the reply looks like a valid message id as
generated by tag_message_id. Since tag values do not currently go
through spamfilter, this is an easy-ish way to ensure that it cannot be
used as a backchannel for spam.

For react, we limit the reaction to a single emoji (potentially
including modifiers) for similar reasons. Additionally, the (currently
unmerged) +draft/unreact was implemented as a means for users to easily
remove their reactions. This dependence on emoji brings in a new
dependency on libicu.